+++ lib/core/models/custom-async-validators.type.ts
@@ -1,0 +1,5 @@
+import { AsyncValidatorFn } from '@angular/forms';
+
+export type CustomAsyncValidators = {
+  [key: string]: AsyncValidatorFn | ((_: any) => AsyncValidatorFn);
+};