+++ lib/core/services/form-generator.service.ts
@@ -1,6 +1,5 @@
 import { Injectable, inject } from '@angular/core';
 import { FormControl, UntypedFormGroup } from '@angular/forms';
-import { ValidatorConfig } from '../models';
 import { FormControlConfig } from '../models/form-control-config.interface';
 import { FormValidationService } from './form-validation.service';
 
@@ -20,7 +19,12 @@ export class FormGeneratorService {
         item.validators
       );
 
+      const asyncValidators = this._formValidationService.getAsyncValidators(
+        item.asyncValidators
+      );
+
       control.setValidators(validators);
+      control.setAsyncValidators(asyncValidators);
       formGroup.addControl(item.formControlName, control);
     }